classification/requirements.txt
google-cloud-vision
pdf2image
google-auth
poppler-utils
python-dotenv
